LayoutTests/http/tests/inspector-enabled/open-close-open.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>

function onload()
{
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
        layoutTestController.showWebInspector();
        layoutTestController.evaluateInWebInspector(0, "true");
        console.log("Opening Web Inspector.");
    }
}

function didEvaluateForTestInFrontend(callId)
{
    if (callId === 0) {
        // Do this asynchronously to allow InspectorBackendDispatcher to send response
        // back to the frontend before it's destroyed.
        setTimeout(function() {
            console.log("Closing Web Inspector.");
            layoutTestController.closeWebInspector();

            console.log("Re-opending Web Inspector.");
            layoutTestController.showWebInspector();

            layoutTestController.evaluateInWebInspector(1, "(" + frontend_testCommunucationWithServer + ")()");
        }, 0);
    }
}

function completeTest()
{
    console.log("Completing the test.");

    // Release 'evaluate' control, report result and on close afterwards.
    setTimeout(function() {
        layoutTestController.closeWebInspector();
        layoutTestController.notifyDone();
    }, 0);
}

function frontend_testCommunucationWithServer()
{
    RuntimeAgent.evaluate("completeTest()", "console", false);
}

</script>

</head>

<body onload="onload()">
<p>
Tests that Web Inspector is functional after the open - close - open sequence.
https://bugs.webkit.org/show_bug.cgi?id=54879
</p>
</body>
</html>
